+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- feat: settlement client introduced instead of just ethereum, starknet client added for settlement
 - fix(gateway-client): fix v0.13.4 gateway deserialization
 - chore: Merge entire madara-orchestrator project into this one
 - fix(primitives): limit legacy class sizes